+++ changelog.md
@@ -5,6 +5,7 @@
 - Adjusted spacing of `\item` content. 
 - Added font size, line height, and alignment config options.
 - Units now get an automatically generated label based on its numbering if none is given. 
+- Added the conservative mode to the `watch` command. 
 
 
 # v0.1.6, 2026-03-19